fix(compat): restore magic-link-only behaviour and extract app factory
Audited against v0.3.1 (main): several gateway-branch changes regressed
the published magic-link-only contract. RUNEGATE_MODE now gates them:

- Public proxy paths (/, favicons, /_app) and the MFA/JWKS surface exist
  only in gateway mode; magic-link-only keeps everything behind auth and
  never takes the PREAUTH -> /mfa branch (redirect-loop guard).
- Login assets resolve to env var -> CWD-relative static/ (the v0.3.x
  location) -> /opt/runegate/static; register page falls back to login
  instead of 500 when register.html is absent.
- Claims.jti is serde(default): in-flight v0.3.x magic links survive an
  upgrade.
- Post-login redirects unified on config::default_redirect(); the
  already-authenticated login/register bounce no longer hardcodes /app.
- Admin handlers take Option<PgStore> and answer 503 instead of a 500
  extraction error when no database is configured.

Route table, UI handlers, and renderer construction move from main.rs to
src/app.rs (library) so integration tests can build the real app
in-process. Deployments with RUNEGATE_MODE=gateway (verbatime/aibox) are
unaffected.
