+++ .env.example
@@ -32,6 +32,49 @@
 # RUNEGATE_MAGIC_LINK_EXPIRY=60
 
 # =============================================================================
+# Operational Mode (see docs/howto/)
+# =============================================================================
+
+# 'magic-link-only' (default) or 'gateway'.
+# magic-link-only preserves the published v0.3.x behaviour: every path except
+# the auth surface requires a session. gateway additionally makes /, the
+# favicons, and /_app/* publicly proxied (for an app with a public landing
+# page) and enables the MFA/JWKS surface.
+# RUNEGATE_MODE=gateway
+
+# Gateway mode: PostgreSQL for durable identity, invites, audit (optional;
+# without it Runegate degrades to stateless behaviour)
+# DATABASE_URL=postgresql://user:pass@localhost:5432/runegate
+
+# Optional: Redis-backed sessions (multi-instance deployments). NOTE: if set
+# and Redis is unreachable, Runegate refuses to start.
+# REDIS_URL=redis://127.0.0.1:6379
+
+# Google OIDC single sign-on (all three required to enable it)
+# RUNEGATE_GOOGLE_CLIENT_ID=your-client-id.apps.googleusercontent.com
+# RUNEGATE_GOOGLE_CLIENT_SECRET=YOUR_CLIENT_SECRET
+# RUNEGATE_GOOGLE_REDIRECT_URL=https://your-domain/auth/google/callback
+
+# Signup policy: 'open' (default) or 'invite_only' (needs DATABASE_URL)
+# RUNEGATE_SIGNUP_POLICY=invite_only
+
+# Bearer token protecting the /admin/invites API (unset = admin API disabled)
+# RUNEGATE_ADMIN_API_TOKEN=YOUR_TOKEN
+
+# Auth UI: 'static' (default), 'phenotyper', or 'external'
+# RUNEGATE_AUTH_UI_MODE=static
+
+# Where the static auth UI assets (login.html, ...) live. Default: the
+# CWD-relative static/ directory if it contains login.html, else
+# /opt/runegate/static.
+# RUNEGATE_LOGIN_ASSETS_DIR=/opt/runegate/static
+
+# Edge authorization: RS256 key pair for minting upload tickets and the JWKS
+# published at /keys/upload_jwks.json (both unset = endpoints answer 501)
+# RUNEGATE_UPLOAD_PRIVATE_KEY=-----BEGIN RSA PRIVATE KEY-----...
+# RUNEGATE_UPLOAD_JWKS={"keys":[...]}
+
+# =============================================================================
 # Logging Configuration
 # =============================================================================
