+++ src/lib.rs
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
+pub mod app;
 pub mod auth;
 pub mod config;
 pub mod email;
