+++ src/main.rs
@@ -1,29 +1,24 @@
 // SPDX-License-Identifier: Apache-2.0
-use actix_files::Files;
-use actix_session::{Session, SessionMiddleware};
+use actix_session::SessionMiddleware;
 use actix_web::cookie::{Key, SameSite};
-use actix_web::http::header;
 use actix_web::middleware::Condition;
-use actix_web::{App, Error, HttpRequest, HttpResponse, HttpServer, Responder, web};
+use actix_web::{App, HttpServer, web};
 
 use std::fs;
 use std::sync::Arc;
 use std::time::Duration;
-use tracing::{debug, error, info, instrument, warn};
+use tracing::{debug, error, info, warn};
 
-use runegate::logging;
 use rand::RngExt;
+use runegate::app::{AppSettings, build_renderer, configure_routes};
+use runegate::config::{AppConfig, AuthUiMode, RunegateMode};
 use runegate::email::EmailConfig;
+use runegate::logging;
 use runegate::memory_session_store::MemorySessionStore;
 use runegate::middleware::AuthMiddleware;
-use runegate::proxy::proxy_request;
 use runegate::rate_limit::RateLimiters;
 use runegate::store::pg::PgStore;
-use tracing_actix_web::TracingLogger; // Added for random key generation
-use runegate::ui::context::{
-    AuthUiContext, RequestContext, MarketContext, LocaleContext,
-    BrandingContext, AuthContext, NavigationContext, SecurityContext, MessageContext
-};
+use tracing_actix_web::TracingLogger;
 
 // Application configuration constants
 const SESSION_KEY_ENV: &str = "RUNEGATE_SESSION_KEY";
@@ -32,257 +27,6 @@ const RUNEGATE_SECURE_COOKIE_VAR: &str = "RUNEGATE_SECURE_COOKIE";
 const RUNEGATE_COOKIE_DOMAIN_VAR: &str = "RUNEGATE_COOKIE_DOMAIN";
 const RUNEGATE_SESSION_COOKIE_NAME_VAR: &str = "RUNEGATE_SESSION_COOKIE_NAME";
 const RUNEGATE_DEBUG_ENDPOINTS_VAR: &str = "RUNEGATE_DEBUG_ENDPOINTS";
-const RUNEGATE_MODE_VAR: &str = "RUNEGATE_MODE";
-const RUNEGATE_AUTH_UI_MODE_VAR: &str = "RUNEGATE_AUTH_UI_MODE";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum RunegateMode {
-    #[default]
-    MagicLinkOnly,
-    Gateway,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum AuthUiMode {
-    #[default]
-    Static,
-    Phenotyper,
-    External,
-}
-
-// We'll get the magic link expiry from environment instead of hardcoding it
-// Default is defined in auth.rs as DEFAULT_MAGIC_LINK_EXPIRY
-
-use runegate::config::AppConfig;
-
-#[derive(serde::Deserialize)]
-struct AuthQueryParams {
-    return_to: Option<String>,
-    locale: Option<String>,
-}
-
-fn create_auth_ui_context(req: &HttpRequest, query: &AuthQueryParams) -> AuthUiContext {
-    let host = req.connection_info().host().to_string();
-    let path = req.path().to_string();
-    let request_id = uuid::Uuid::new_v4().to_string();
-
-    let return_to = query.return_to.clone().unwrap_or_else(|| "/app".to_string());
-    let locale_str = query.locale.clone().unwrap_or_else(|| "en".to_string());
-
-    AuthUiContext {
-        request: RequestContext {
-            host,
-            path,
-            request_id,
-        },
-        market: MarketContext {
-            id: "global".to_string(),
-            canonical_domain: "verbatime.ai".to_string(),
-            market_name: "Global".to_string(),
-            currency: "EUR".to_string(),
-        },
-        locale: LocaleContext {
-            current: locale_str,
-            available: vec!["en".to_string(), "nl".to_string(), "nl-BE".to_string(), "de".to_string(), "fr".to_string()],
-            default: "en".to_string(),
-        },
-        branding: BrandingContext {
-            product_name: "Verbatime".to_string(),
-            logo_url: "/img/logo.svg".to_string(),
-            primary_color: "#0ea5e9".to_string(),
-            support_email: "support@verbatime.dev".to_string(),
-        },
-        auth: AuthContext {
-            mode: "gateway".to_string(),
-            signup_policy: "open".to_string(),
-            magic_link_enabled: true,
-            google_enabled: true,
-            webauthn_enabled: true,
-            totp_enabled: true,
-        },
-        navigation: NavigationContext {
-            return_to,
-            login_url: "/auth/login".to_string(),
-            register_url: "/auth/register".to_string(),
-        },
-        security: SecurityContext {
-            csrf_token: "".to_string(),
-            csp_nonce: "".to_string(),
-        },
-        messages: MessageContext {
-            title: "".to_string(),
-            subtitle: None,
-            flash: None,
-            error: None,
-        },
-    }
-}
-
-async fn serve_login(
-    req: HttpRequest,
-    query: web::Query<AuthQueryParams>,
-    renderer: web::Data<dyn runegate::ui::AuthUiRenderer>,
-    session: Session,
-) -> impl Responder {
-    let is_authenticated = session
-        .get::<bool>("authenticated")
-        .unwrap_or(None)
-        .unwrap_or(false);
-    if is_authenticated {
-        let return_to = query.return_to.clone().unwrap_or_else(|| "/app".to_string());
-        return HttpResponse::Found()
-            .append_header((header::LOCATION, return_to))
-            .finish();
-    }
-
-    let ctx = create_auth_ui_context(&req, &query);
-    match renderer.render_login(&ctx) {
-        Ok(resp) => resp,
-        Err(e) => {
-            error!("Failed to render login UI: {}", e);
-            HttpResponse::InternalServerError().body(format!("Failed to render login UI: {}", e))
-        }
-    }
-}
-
-async fn serve_register(
-    req: HttpRequest,
-    query: web::Query<AuthQueryParams>,
-    renderer: web::Data<dyn runegate::ui::AuthUiRenderer>,
-    session: Session,
-) -> impl Responder {
-    let is_authenticated = session
-        .get::<bool>("authenticated")
-        .unwrap_or(None)
-        .unwrap_or(false);
-    if is_authenticated {
-        let return_to = query.return_to.clone().unwrap_or_else(|| "/app".to_string());
-        return HttpResponse::Found()
-            .append_header((header::LOCATION, return_to))
-            .finish();
-    }
-
-    let ctx = create_auth_ui_context(&req, &query);
-    match renderer.render_register(&ctx) {
-        Ok(resp) => resp,
-        Err(e) => {
-            error!("Failed to render register UI: {}", e);
-            HttpResponse::InternalServerError().body(format!("Failed to render register UI: {}", e))
-        }
-    }
-}
-
-async fn serve_mfa_select(
-    req: HttpRequest,
-    query: web::Query<AuthQueryParams>,
-    renderer: web::Data<dyn runegate::ui::AuthUiRenderer>,
-) -> impl Responder {
-    let ctx = create_auth_ui_context(&req, &query);
-    match renderer.render_mfa_select(&ctx) {
-        Ok(resp) => resp,
-        Err(e) => {
-            error!("Failed to render MFA select UI: {}", e);
-            HttpResponse::InternalServerError().body(format!("Failed to render MFA select: {}", e))
-        }
-    }
-}
-
-/// Health check endpoint
-#[instrument(name = "health_check", skip_all)]
-async fn health_check() -> impl Responder {
-    let version = env!("CARGO_PKG_VERSION");
-    HttpResponse::Ok().json(serde_json::json!({
-        "status": "running",
-        "service": "Runegate",
-        "version": version
-    }))
-}
-
-/// Authentication check and proxy handler
-#[instrument(name = "auth_check_and_proxy", skip(payload, session), fields(path = %req.path(), method = %req.method()))]
-async fn auth_check_and_proxy(
-    req: HttpRequest,
-    payload: web::Payload,
-    session: Session,
-) -> Result<HttpResponse, Error> {
-    // Check if user is authenticated
-    debug!(
-        "[PROXY_AUTH - EVENT] Checking session for proxy request to: {}",
-        req.path()
-    );
-
-    match session.get::<bool>("authenticated") {
-        Ok(Some(val)) => debug!(
-            "[PROXY_AUTH - EVENT] Session authenticated result: Ok(Some({}))",
-            val
-        ),
-        Ok(None) => debug!("[PROXY_AUTH - EVENT] Session authenticated result: Ok(None)"),
-        Err(e) => debug!("[PROXY_AUTH - EVENT] Session authenticated error: {}", e),
-    }
-
-    match session.get::<String>("email") {
-        Ok(Some(val)) => debug!(
-            "[PROXY_AUTH - EVENT] Session email result: Ok(Some({}))",
-            val
-        ),
-        Ok(None) => debug!("[PROXY_AUTH - EVENT] Session email result: Ok(None)"),
-        Err(e) => debug!("[PROXY_AUTH - EVENT] Session email error: {}", e),
-    }
-
-    let is_authenticated = session
-        .get::<bool>("authenticated")
-        .unwrap_or(None)
-        .unwrap_or(false);
-    debug!(
-        "[PROXY_AUTH - EVENT] Final authenticated value: {}",
-        is_authenticated
-    );
-
-    let path = req.path();
-    let is_public_path = path == "/"
-        || path == "/favicon.ico"
-        || path == "/favicon.svg"
-        || path.starts_with("/_app");
-
-    if is_authenticated || is_public_path {
-        // User is authenticated, or this is a public proxy path (e.g. landing page or assets).
-        // Proxy the request. If authenticated, inject identity headers.
-        let identity_email = if is_authenticated {
-            session.get::<String>("email").ok().flatten()
-        } else {
-            None
-        };
-        proxy_request(req, payload, identity_email).await
-    } else {
-        // User is not authenticated, redirect to login
-        // Detect if we're behind a proxy and construct the correct redirect path
-        let redirect_path = if req.headers().contains_key("X-Forwarded-Proto") {
-            // We're behind a proxy, need to determine the base path
-            let original_uri = req.uri().path();
-            if original_uri == "/" {
-                // We're at the root of the proxied path, redirect to login.html at the same level
-                "./login.html".to_string()
-            } else {
-                // Extract the base path from the original URI
-                let path_segments: Vec<&str> =
-                    original_uri.trim_start_matches('/').split('/').collect();
-                if path_segments.len() > 1 {
-                    format!("/{}/login.html", path_segments[0])
-                } else {
-                    "/login.html".to_string()
-                }
-            }
-        } else {
-            // Direct access, use absolute path
-            "/login.html".to_string()
-        };
-
-        debug!("Redirecting unauthenticated user to: {}", redirect_path);
-        Ok(HttpResponse::Found()
-            .append_header((header::LOCATION, redirect_path))
-            .finish())
-    }
-}
 
 /// Load configuration from TOML file
 fn load_config() -> AppConfig {
@@ -429,12 +173,13 @@ fn log_environment_config() {
     info!("🔧 Environment mode: {}", env_mode);
 
     // Runegate mode (magic-link-only or gateway)
-    let runegate_mode =
-        std::env::var(RUNEGATE_MODE_VAR).unwrap_or_else(|_| "magic-link-only".to_string());
+    let runegate_mode = std::env::var(runegate::config::RUNEGATE_MODE_ENV)
+        .unwrap_or_else(|_| "magic-link-only".to_string());
     info!("🔄 Runegate mode: {}", runegate_mode);
 
     // Auth UI mode (static, phenotyper, external)
-    let ui_mode = std::env::var(RUNEGATE_AUTH_UI_MODE_VAR).unwrap_or_else(|_| "static".to_string());
+    let ui_mode = std::env::var(runegate::config::RUNEGATE_AUTH_UI_MODE_ENV)
+        .unwrap_or_else(|_| "static".to_string());
     info!("🎨 Auth UI mode: {}", ui_mode);
 
     // JWT Secret (length only for security)
@@ -537,99 +282,6 @@ fn load_env_file() {
     );
 }
 
-/// Diagnostic endpoint to return the current rate limiting configuration
-#[instrument(name = "rate_limit_info")]
-async fn rate_limit_info(rate_limiters: web::Data<Arc<RateLimiters>>) -> impl Responder {
-    let rate_limit_config = rate_limiters.config.clone();
-    HttpResponse::Ok().json(rate_limit_config)
-}
-
-/// Debug endpoint to inspect server-side session view
-#[instrument(name = "debug_session", skip(session, req))]
-async fn debug_session(req: HttpRequest, session: Session) -> impl Responder {
-    let session_status = format!("{:?}", session.status());
-    let entries = session.entries();
-    let entry_keys: Vec<String> = entries.keys().cloned().collect();
-    let authenticated = session.get::<bool>("authenticated").ok().flatten();
-    let email = session.get::<String>("email").ok().flatten();
-    let cookie_header = req
-        .headers()
-        .get(header::COOKIE)
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string());
-    let client_ip = req
-        .connection_info()
-        .realip_remote_addr()
-        .unwrap_or("unknown")
-        .to_string();
-    let pid = std::process::id();
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "session_status": session_status,
-        "entry_keys": entry_keys,
-        "authenticated": authenticated,
-        "email": email,
-        "cookie_header": cookie_header,
-        "client_ip": client_ip,
-        "pid": pid,
-    }))
-}
-
-/// Debug endpoint to inspect parsed cookies
-#[instrument(name = "debug_cookies", skip(req))]
-async fn debug_cookies(req: HttpRequest) -> impl Responder {
-    let raw_cookie_header = req
-        .headers()
-        .get(header::COOKIE)
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string());
-    let parsed = match req.cookies() {
-        Ok(cs) => Some(
-            cs.iter()
-                .map(|c| {
-                    serde_json::json!({
-                        "name": c.name(),
-                        "value": c.value(),
-                        "path": c.path().map(|s| s.to_string()),
-                        "domain": c.domain().map(|s| s.to_string()),
-                        "http_only": c.http_only().unwrap_or(false),
-                        "secure": c.secure().unwrap_or(false),
-                        "same_site": c.same_site().map(|s| format!("{:?}", s)),
-                    })
-                })
-                .collect::<Vec<_>>(),
-        ),
-        Err(_e) => None,
-    };
-    let client_ip = req
-        .connection_info()
-        .realip_remote_addr()
-        .unwrap_or("unknown")
-        .to_string();
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "raw_cookie_header": raw_cookie_header,
-        "parsed": parsed,
-        "client_ip": client_ip,
-    }))
-}
-
-/// Debug endpoint that goes through auth middleware to verify auth gating
-#[instrument(name = "debug_protected", skip(session))]
-async fn debug_protected(session: Session) -> impl Responder {
-    let authenticated = session
-        .get::<bool>("authenticated")
-        .ok()
-        .flatten()
-        .unwrap_or(false);
-    let email = session.get::<String>("email").ok().flatten();
-    HttpResponse::Ok().json(serde_json::json!({
-        "authenticated": authenticated,
-        "email": email,
-        "note": "This endpoint requires auth via middleware. Redirects to /login.html if not authed.",
-    }))
-}
-
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
     // Load .env file from multiple possible locations
@@ -658,38 +310,14 @@ async fn main() -> std::io::Result<()> {
     let version = env!("CARGO_PKG_VERSION");
     info!("🚪 Starting Runegate auth proxy v{}", version);
 
-    // Parse RUNEGATE_MODE
-    let runegate_mode = match std::env::var(RUNEGATE_MODE_VAR).as_deref() {
-        Ok("gateway") => RunegateMode::Gateway,
-        _ => RunegateMode::MagicLinkOnly,
-    };
+    let runegate_mode = RunegateMode::from_env();
     info!("🚀 Operating in {:?} mode", runegate_mode);
 
-    // Parse RUNEGATE_AUTH_UI_MODE
-    let auth_ui_mode = match std::env::var(RUNEGATE_AUTH_UI_MODE_VAR).as_deref() {
-        Ok("phenotyper") => AuthUiMode::Phenotyper,
-        Ok("external") => AuthUiMode::External,
-        _ => AuthUiMode::Static,
-    };
+    let auth_ui_mode = AuthUiMode::from_env();
     info!("🎨 Auth UI configured for {:?} mode", auth_ui_mode);
 
     // Initialize Auth UI Renderer
-    let renderer: std::sync::Arc<dyn runegate::ui::AuthUiRenderer> = match auth_ui_mode {
-        AuthUiMode::Phenotyper => {
-            let template_dir = std::env::var("RUNEGATE_TEMPLATE_DIR")
-                .unwrap_or_else(|_| "/opt/runegate/templates".to_string());
-            std::sync::Arc::new(runegate::ui::phenotyper_renderer::PhenotyperRenderer::new(
-                std::path::PathBuf::from(template_dir),
-            ))
-        }
-        AuthUiMode::Static | AuthUiMode::External => {
-            let assets_dir = std::env::var("RUNEGATE_LOGIN_ASSETS_DIR")
-                .unwrap_or_else(|_| "/opt/runegate/static".to_string());
-            std::sync::Arc::new(runegate::ui::static_renderer::StaticRenderer::new(
-                std::path::PathBuf::from(assets_dir),
-            ))
-        }
-    };
+    let renderer = build_renderer(auth_ui_mode);
     let renderer_data = web::Data::from(renderer);
 
     // Log environment configuration (redacting sensitive values)
@@ -745,6 +373,12 @@ async fn main() -> std::io::Result<()> {
         info!("🧪 Debug endpoints DISABLED");
     }
 
+    let app_settings = AppSettings {
+        mode: runegate_mode,
+        debug_endpoints_enabled,
+    };
+    let app_settings_data = web::Data::new(app_settings);
+
     // Determine worker count from environment (default to 2)
     let workers_env = std::env::var("RUNEGATE_WORKERS").ok();
     let workers = workers_env
@@ -762,7 +396,7 @@ async fn main() -> std::io::Result<()> {
 
     HttpServer::new(move || {
         let pg_store_instance = pg_store.clone();
-        
+
         // Determine cookie_secure setting
         let secure_cookie = match std::env::var(RUNEGATE_SECURE_COOKIE_VAR).as_deref() {
             Ok("true") => {
@@ -800,78 +434,32 @@ async fn main() -> std::io::Result<()> {
             None => info!("No cookie domain set; using host-only cookies"),
         }
 
-        {
-            let mut app = App::new()
-                // Access logging (gate by env; default off in production)
-                .wrap(Condition::new(request_logs_enabled, TracingLogger::default()))
-                // Ensure SessionMiddleware runs before AuthMiddleware so session is available in auth checks
-                .wrap(AuthMiddleware::new())
-                .wrap(
-                    SessionMiddleware::builder(shared_session_store.clone(), session_key.clone())
-                        .cookie_secure(secure_cookie)
-                        .cookie_http_only(true)
-                        .cookie_same_site(SameSite::Lax)
-                        .cookie_path("/".to_string())
-                        .cookie_domain(cookie_domain_opt)
-                        .cookie_name(session_cookie_name.clone())
-                        .build()
-                )
-                // App data
-                .app_data(app_config.clone())
-                .app_data(rate_limiters_data.clone())
-                .app_data(renderer_data.clone());
-                
-            if let Some(pg) = &pg_store_instance {
-                app = app.app_data(web::Data::new(pg.clone()));
-            }
+        let mut app = App::new()
+            // Access logging (gate by env; default off in production)
+            .wrap(Condition::new(request_logs_enabled, TracingLogger::default()))
+            // Ensure SessionMiddleware runs before AuthMiddleware so session is available in auth checks
+            .wrap(AuthMiddleware::with_mode(app_settings.mode))
+            .wrap(
+                SessionMiddleware::builder(shared_session_store.clone(), session_key.clone())
+                    .cookie_secure(secure_cookie)
+                    .cookie_http_only(true)
+                    .cookie_same_site(SameSite::Lax)
+                    .cookie_path("/".to_string())
+                    .cookie_domain(cookie_domain_opt)
+                    .cookie_name(session_cookie_name.clone())
+                    .build(),
+            )
+            // App data
+            .app_data(app_config.clone())
+            .app_data(rate_limiters_data.clone())
+            .app_data(renderer_data.clone())
+            .app_data(app_settings_data.clone());
 
-            let mut app = app
-                // API Endpoints - define these first to ensure they take priority
-                .service(web::resource("/health").route(web::get().to(health_check)))
-                .service(web::resource("/auth/identify").route(web::post().to(runegate::routes::auth::identify)))
-                .service(web::resource("/auth/magic/start").route(web::post().to(runegate::routes::auth::magic_start)))
-                .service(web::resource("/auth/magic/consume").route(web::get().to(runegate::routes::auth::magic_consume)))
-                // Aliases for backward compatibility
-                .service(web::resource("/login").route(web::post().to(runegate::routes::auth::magic_start)))
-                .service(web::resource("/auth").route(web::get().to(runegate::routes::auth::magic_consume)))
-                // OIDC Endpoints
-                .service(web::resource("/auth/google/start").route(web::get().to(runegate::routes::oidc::google_start)))
-                .service(web::resource("/auth/google/callback").route(web::get().to(runegate::routes::oidc::google_callback)))
-                // MFA Endpoints
-                .service(web::resource("/mfa/totp/verify").route(web::post().to(runegate::routes::mfa::totp_verify)))
-                .service(web::resource("/mfa/webauthn/start").route(web::post().to(runegate::routes::mfa::webauthn_start)))
-                .service(web::resource("/mfa/webauthn/finish").route(web::post().to(runegate::routes::mfa::webauthn_finish)))
-                // Upload Endpoints
-                .service(web::resource("/upload-ticket").route(web::post().to(runegate::routes::upload::create_upload_ticket)))
-                .service(web::resource("/keys/upload_jwks.json").route(web::get().to(runegate::routes::upload::get_upload_jwks)))
-                // Admin Endpoints
-                .service(web::resource("/admin/invites").route(web::post().to(runegate::routes::admin::create_invite)).route(web::get().to(runegate::routes::admin::get_invites)))
-                .service(web::resource("/admin/invites/{id}/revoke").route(web::post().to(runegate::routes::admin::revoke_invite)))
-                .service(web::resource("/rate_limit_info").route(web::get().to(rate_limit_info)))
-                // UI Endpoints
-                .service(web::resource("/auth/login").route(web::get().to(serve_login)))
-                .service(web::resource("/auth/login.html").route(web::get().to(serve_login)))
-                .service(web::resource("/login.html").route(web::get().to(serve_login)))
-                .service(web::resource("/auth/register").route(web::get().to(serve_register)))
-                .service(web::resource("/auth/register.html").route(web::get().to(serve_register)))
-                .service(web::resource("/register.html").route(web::get().to(serve_register)))
-                .service(web::resource("/mfa").route(web::get().to(serve_mfa_select)));
-
-            if debug_endpoints_enabled {
-                app = app
-                    .service(web::resource("/debug/session").route(web::get().to(debug_session)))
-                    .service(web::resource("/debug/cookies").route(web::get().to(debug_cookies)))
-                    .service(web::resource("/debug/protected").route(web::get().to(debug_protected)));
-            }
-
-            app
-                // Static files serving - place after API endpoints to avoid routing conflicts
-                .service(Files::new("/auth", "static"))
-                .service(Files::new("/static", "static"))
-                .service(Files::new("/img", "static/img"))
-                // Protected routes need to be guarded in each handler
-                .default_service(web::route().to(auth_check_and_proxy))
+        if let Some(pg) = &pg_store_instance {
+            app = app.app_data(web::Data::new(pg.clone()));
         }
+
+        app.configure(|cfg| configure_routes(cfg, &app_settings))
     })
     .bind("0.0.0.0:7870")?
     .client_request_timeout(Duration::from_secs(600))
